Fix problem with adding items to database.
Caused by the clicked() signals of the relevant Store buttons on the details dialogs being connected to non-implemented slots in JpCnPluginBase. Solved by reconnecting to MainForm::slot_dict_toDbaseClicked() instead, it works by extracting the current item from the table but since the table item has to exist and be selected while the details dialogs are popped, it doesn't make any difference and saves the effort of scraping the item/desc data from the details dialog contents.