Fixed problem with crash on copy/move.
Caused by lack of check if pointer for target database was valid before attempting the operation. Since the handle is extracted based on the string in the combobox which was empty, this returned a null pointer which was later dereferenced leading to the crash.

In any case, also introduced a check to see if the list of target databases on the dialog is empty and if so, don't pop up the dialog at all, show an information message box instead that the operation cannot be performed.